medvision/requirements.txt
google-generativeai
streamlit
python-docx
Pillow
PyMuPDF
python-dotenv